field sizes set, fixed blank fields exceptions, added description book, unlimited random books, automated field available depending on the field count

+++ src/main/java/pl/library/domain/book/repository/BookRepository.java
@@ -12,7 +12,7 @@ public interface BookRepository extends JpaRepository<Book, Long> {
     @Query(value = "SELECT * FROM books WHERE title LIKE %:phrase% or author LIKE %:phrase% ORDER BY title ASC", nativeQuery = true)
     List<Book> searchAllByTitleOrAuthorLike(@Param("phrase") String phrase);
     @Query(value = "SELECT * FROM books ORDER BY RAND() LIMIT :number", nativeQuery = true)
-    List<Book> searchNBooksByRandom(@Param("number") Integer number);
+    List<Book> searchNBooksByRandom(@Param("number") Byte number);
     Optional<Book> findByIdAndTitle(Long id, String title);
 //    List<Book> findAllByGenres(String genre);
     Boolean existsByTitleAndAuthor(String title, String author);
